// src/components/types/accounts/accountLogin.dto.ts
export class AccountLoginDto {
  id: number;
  name: string;
  password: string;
  email: string;
  premdays: number;
  premdays_purchased: number;
  lastday: number;
  type: number;
  coins: number;
  coins_transferable: number;
  tournament_coins: number;
  creation: number;
  recruiter: number;
}